Store the previous PC when a PM IRQ occurs
This way, the normal control flow access control logic can be used when
a RETI to PM is executed. Before, we used to store a single bit to
indicate that the previous instruction was a RETI to PM and the access
control logic would take that into account to allow jumps to non-entry
points.  However, this scheme is only secure with a single module. Also,
storing the previous PC more closely follows the model.